+++ src/main/java/dev/ftb/mods/ftbic/net/FTBICNet.java
@@ -15,6 +15,8 @@ public final class FTBICNet {
 	@SubscribeEvent
 	public static void registerPayloads(RegisterPayloadHandlersEvent event) {
 		var registrar = event.registrar(FTBIC.MOD_ID);
+		registrar.playToServer(BatchFluidPayload.TYPE, BatchFluidPayload.STREAM_CODEC, BatchFluidPayload::handleOnServer);
+		registrar.playToServer(GhostSlotPayload.TYPE, GhostSlotPayload.STREAM_CODEC, GhostSlotPayload::handleOnServer);
 		registrar.playToServer(SideConfigurationPayload.TYPE, SideConfigurationPayload.STREAM_CODEC, SideConfigurationPayload::handleOnServer);
 		registrar.playToServer(ReactorDesignPayload.TYPE, ReactorDesignPayload.STREAM_CODEC, ReactorDesignPayload::handleOnServer);